Replace Entrypoint module with ApiController
* pulls this code out of the Api::BaseController
* allows for the more natural `get "/api(/:version)" => "api#index"`
